Adding unit tests for GDPR Factory class
